requirements.txt
Flask==3.1.0
python-dotenv==1.1.0
spotipy==2.25.1